+++ apps/web/src/views/TradingReward/components/YourTradingReward/index.tsx
@@ -4,9 +4,9 @@ import BigNumber from 'bignumber.js'
 import { useMemo } from 'react'
 import { useProfile } from 'state/profile/hooks'
 import { styled } from 'styled-components'
+import { useVeCakeUserCreditWithTime } from 'views/CakeStaking/hooks/useVeCakeUserCreditWithTime'
 import { useCakeLockStatus } from 'views/CakeStaking/hooks/useVeCakeUserInfo'
 import { floatingStarsLeft, floatingStarsRight } from 'views/Lottery/components/Hero'
-import { useVeCakeUserCreditWithTime } from 'views/Pools/hooks/useVeCakeUserCreditWithTime'
 import NoConnected from 'views/TradingReward/components/YourTradingReward/NoConnected'
 import NoProfile from 'views/TradingReward/components/YourTradingReward/NoProfile'
 import RewardPeriod from 'views/TradingReward/components/YourTradingReward/RewardPeriod'
