+++ packages/ifos/src/index.test.ts
@@ -38,7 +38,6 @@ test('exports', () => {
       "fetchUserVestingData",
       "getBridgeICakeGasFee",
       "getCrossChainMessage",
-      "getUserCreditWithTime,
       "iCakeABI",
       "pancakeInfoSenderABI",
       "ifoV7ABI",
