+++ packages/ifos/src/queries/index.ts
@@ -1,4 +1,3 @@
 export * from './fetchUserIfo'
 export * from './getBridgeICakeGasFee'
 export * from './getCrossChainMessage'
-export * from './getUserCreditWithTime'
